# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "safehaven",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin/access_tokens": {
      "get": {
        "tags": [
          "admin::access_tokens"
        ],
        "operationId": "admin_access_tokens_list",
        "responses": {
          "200": {
            "description": "List of access tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccessToken"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin::access_tokens"
        ],
        "operationId": "admin_access_token_new",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewOrUpdateAccessToken"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "List of access tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/access_tokens/{id}": {
      "get": {
        "tags": [
          "admin::access_tokens"
        ],
        "operationId": "admin_access_token_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Access token identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin::access_tokens"
        ],
        "operationId": "admin_access_token_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Access token identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewOrUpdateAccessToken"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin::access_tokens"
        ],
        "operationId": "admin_access_token_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Access token identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion successful"
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/categories": {
      "get": {
        "tags": [
          "admin::categories"
        ],
        "operationId": "admin_categories_list",
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin::categories"
        ],
        "operationId": "admin_category_new",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCategory"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/categories/{id}": {
      "get": {
        "tags": [
          "admin::categories"
        ],
        "operationId": "admin_category_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Category identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin::categories"
        ],
        "operationId": "admin_category_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Category identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategory"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Category updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin::categories"
        ],
        "operationId": "admin_category_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Category identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deletion successful"
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessToken": {
        "type": "object",
        "required": [
          "id",
          "token",
          "permissions",
          "active"
        ],
        "properties": {
          "active": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "permissions": {
            "$ref": "#/components/schemas/Json"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "Category": {
        "type": "object",
        "required": [
          "id",
          "title",
          "family_id",
          "default_status",
          "fill_color",
          "border_color"
        ],
        "properties": {
          "border_color": {
            "type": "string"
          },
          "default_status": {
            "type": "boolean"
          },
          "family_id": {
            "type": "string",
            "format": "uuid"
          },
          "fill_color": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error_code"
        ],
        "properties": {
          "details": {
            "type": "string",
            "nullable": true
          },
          "error_code": {
            "type": "string"
          }
        }
      },
...
      "NewCategory": {
        "type": "object",
        "required": [
          "title",
          "family_id"
        ],
        "properties": {
          "family_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "NewOrUpdateAccessToken": {
        "type": "object",
        "required": [
          "token",
          "permissions",
          "active"
        ],
        "properties": {
          "active": {
            "type": "boolean"
          },
          "permissions": {
            "$ref": "#/components/schemas/Json"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "UpdateCategory": {
        "type": "object",
        "required": [
          "title",
          "family_id",
          "default_status",
          "fill_color",
          "border_color"
        ],
        "properties": {
          "border_color": {
            "type": "string"
          },
          "default_status": {
            "type": "boolean"
          },
          "family_id": {
            "type": "string",
            "format": "uuid"
          },
          "fill_color": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string"
          }
        }
      },
...
    }
  }
}